Following a tutorial on Pygame basics on youtube - it has been good.
I've been able to make it to the game state changes section and have
finally found some success.

I've found a way to handle events in a more organized way; so far so good.

After rewriting a days worth of code because of mixing up code, it is back to where I can make progress.

All code content needed should be upload now.